fix price type

+++ HomeCook/Models/Dish.swift
@@ -2,8 +2,8 @@
 struct Dish {
     let title: String
     let ingredients: [Ingredient]
-    var price: Double {
-        var cost = 0.0
+    var price: Int {
+        var cost = 0
         ingredients.forEach { ingredient in cost += ingredient.price }
         return cost
     }
